// lib/Api/api.ts
export interface LoginUser {
    /**
     * 
     * @type {string}
     * @memberof LoginUser
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof LoginUser
     */
    password: string;
}
/**
 * 
 * @export
 * @interface LoginUserRequest
 */
export interface LoginUserRequest {
    /**
     * 
     * @type {LoginUser}
     * @memberof LoginUserRequest
     */
    user: LoginUser;
}
// ...
export interface NewUser {
    /**
     * 
     * @type {string}
     * @memberof NewUser
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof NewUser
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof NewUser
     */
    password: string;
}
/**
 * 
 * @export
 * @interface NewUserRequest
 */
export interface NewUserRequest {
    /**
     * 
     * @type {NewUser}
     * @memberof NewUserRequest
     */
    user: NewUser;
}
// ...
export interface UpdateUser {
    /**
     * 
     * @type {string}
     * @memberof UpdateUser
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUser
     */
    token?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUser
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUser
     */
    bio?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUser
     */
    image?: string;
}
/**
 * 
 * @export
 * @interface UpdateUserRequest
 */
export interface UpdateUserRequest {
    /**
     * 
     * @type {UpdateUser}
     * @memberof UpdateUserRequest
     */
    user: UpdateUser;
}
// ...
export class UserAndAuthenticationApi extends BaseAPI {
    /**
     * Register a new user
     * @summary Register a new user
     * @param {NewUserRequest} body Details of the new user to register
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserAndAuthenticationApi
     */
    public createUser(body: NewUserRequest, options?: any) {
        return UserAndAuthenticationApiFp(this.configuration).createUser(body, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Gets the currently logged-in user
     * @summary Get current user
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserAndAuthenticationApi
     */
    public getCurrentUser(options?: any) {
        return UserAndAuthenticationApiFp(this.configuration).getCurrentUser(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Login for existing user
     * @summary Existing user login
     * @param {LoginUserRequest} body Credentials to use
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserAndAuthenticationApi
     */
    public login(body: LoginUserRequest, options?: any) {
        return UserAndAuthenticationApiFp(this.configuration).login(body, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Updated user information for current user
     * @summary Update current user
     * @param {UpdateUserRequest} body User details to update. At least **one** field is required.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserAndAuthenticationApi
     */
    public updateCurrentUser(body: UpdateUserRequest, options?: any) {
        return UserAndAuthenticationApiFp(this.configuration).updateCurrentUser(body, options).then((request) => request(this.axios, this.basePath));
    }
}